feat: Implement dry-run mode for `session-manager port-forward` command

+++ aws_annoying/cli/session_manager/port_forward.py
@@ -19,6 +19,8 @@ logger = logging.getLogger(__name__)
 # https://docs.aws.amazon.com/systems-manager/latest/userguide/session-manager-working-with-install-plugin.html
 @session_manager_app.command()
 def port_forward(  # noqa: PLR0913
+    ctx: typer.Context,
+    *,
     # TODO(lasuillard): Add `--local-host` option, redirect the traffic to non-localhost bind (unsupported by AWS)
     local_port: int = typer.Option(
         ...,
@@ -48,7 +50,7 @@ def port_forward(  # noqa: PLR0913
         "./session-manager-plugin.pid",
         help="The path to the PID file to store the process ID of the session manager plugin.",
     ),
-    terminate_running_process: bool = typer.Option(  # noqa: FBT001
+    terminate_running_process: bool = typer.Option(
         False,  # noqa: FBT003
         help="Terminate the process in the PID file if it already exists.",
     ),
@@ -58,6 +60,7 @@ def port_forward(  # noqa: PLR0913
     ),
 ) -> None:
     """Start a port forwarding session using AWS Session Manager."""
+    dry_run = ctx.meta["dry_run"]
     session_manager = SessionManager()
 
     # Check if the PID file already exists
@@ -111,6 +114,10 @@ def port_forward(  # noqa: PLR0913
         through,
         reason,
     )
+    if dry_run:
+        # If dry run mode is enabled, just print the command (the log file will catch this)
+        command = ["echo", *command]
+
     proc = subprocess.Popen(  # noqa: S603
         command,
         stdout=stdout,
@@ -125,5 +132,7 @@ def port_forward(  # noqa: PLR0913
     )
 
     # Write the PID to the file
-    pid_file.write_text(str(proc.pid))
+    if not dry_run:
+        pid_file.write_text(str(proc.pid))
+
     logger.info("PID file written to %s.", pid_file.absolute())
